+++ src/cli/app/app-list.js
@@ -1,14 +1,14 @@
-import { Command, Option } from 'commander';
+import { Command } from 'commander';
 import * as common from '../cmd_common.js';
 import { getTokens } from '../../ops/AuthenticateOps.js';
 import storage from '../../storage/SessionStorage.js';
 import { printMessage } from '../../ops/utils/Console.js';
-import { listScripts } from '../../ops/ScriptOps.js';
+import { listOAuth2Applications } from '../../api/ApplicationApi.js';
 
-const program = new Command('frodo script list');
+const program = new Command('frodo app list');
 
 program
-  .description('List scripts.')
+  .description('List OAuth2 applications.')
   .helpOption('-h, --help', 'Help')
   .showHelpAfterError()
   .addArgument(common.hostArgumentM)
@@ -17,9 +17,9 @@ program
   .addArgument(common.passwordArgument)
   .addOption(common.deploymentOption)
   .addOption(common.insecureOption)
-  .addOption(
-    new Option('-l, --long', 'Long with all fields.').default(false, 'false')
-  )
+  // .addOption(
+  //   new Option('-l, --long', 'Long with all fields.').default(false, 'false')
+  // )
   .action(
     // implement command logic inside action handler
     async (host, realm, user, password, options) => {
@@ -30,10 +30,12 @@ program
       storage.session.setDeploymentType(options.type);
       storage.session.setAllowInsecureConnection(options.insecure);
       if (await getTokens()) {
-        printMessage(
-          `Listing scripts in realm "${storage.session.getRealm()}"...`
-        );
-        await listScripts(options.long);
+        printMessage(`Listing OAuth2 applications...`);
+        const applicationList = await listOAuth2Applications();
+        applicationList.sort((a, b) => a._id.localeCompare(b._id));
+        applicationList.forEach((item) => {
+          printMessage(`${item._id}`);
+        });
       }
     }
     // end command logic inside action handler