+++ src/cli/realm/realm-list.js
@@ -2,13 +2,13 @@ import { Command, Option } from 'commander';
 import * as common from '../cmd_common.js';
 import { getTokens } from '../../ops/AuthenticateOps.js';
 import storage from '../../storage/SessionStorage.js';
-import { printMessage } from '../../ops/utils/Console.js';
-import { listScripts } from '../../ops/ScriptOps.js';
+import { createTable, printMessage } from '../../ops/utils/Console.js';
+import { listRealms } from '../../api/RealmApi.js';
 
-const program = new Command('frodo script list');
+const program = new Command('frodo realm list');
 
 program
-  .description('List scripts.')
+  .description('List realms.')
   .helpOption('-h, --help', 'Help')
   .showHelpAfterError()
   .addArgument(common.hostArgumentM)
@@ -30,10 +30,29 @@ program
       storage.session.setDeploymentType(options.type);
       storage.session.setAllowInsecureConnection(options.insecure);
       if (await getTokens()) {
-        printMessage(
-          `Listing scripts in realm "${storage.session.getRealm()}"...`
-        );
-        await listScripts(options.long);
+        printMessage('Listing all realms...');
+        const realms = await listRealms();
+        if (options.long) {
+          const table = createTable([
+            'Name'.brightCyan,
+            'Status'.brightCyan,
+            'Custom Domains'.brightCyan,
+            'Parent'.brightCyan,
+          ]);
+          realms.forEach((realmConfig) => {
+            table.push([
+              realmConfig.name,
+              realmConfig.active ? 'active'.brightGreen : 'inactive'.brightRed,
+              realmConfig.aliases.join('\n'),
+              realmConfig.parentPath,
+            ]);
+          });
+          printMessage(table.toString());
+        } else {
+          realms.forEach((realmConfig) => {
+            printMessage(realmConfig.name, 'info');
+          });
+        }
       }
     }
     // end command logic inside action handler